+++ worker/jobs/pull/__init__test.py
@@ -3,15 +3,15 @@ import pytest
 from . import Pull
 
 
-def test_missing_provider():
+def test_missing_type():
     pull = Pull()
     with pytest.raises(AssertionError) as excinfo:
-        pull.do({}, "sink")
-    assert "Source must have a provider" in str(excinfo.value)
+        pull.do({}, 1, "path")
+    assert "source must have a type" in str(excinfo.value)
 
 
-def test_unknown_provider():
+def test_unknown_type():
     pull = Pull()
     with pytest.raises(ValueError) as excinfo:
-        pull.do({"provider": "foo"}, "sink")
-    assert "Unknown source provider: foo" in str(excinfo.value)
+        pull.do({"type": "foo"}, 1, "path")
+    assert "Unknown source type: foo" in str(excinfo.value)