+++ worker/jobs/pull/plos.py
@@ -1,13 +1,14 @@
 import io
 import os
 import re
+from typing import List
 
 from lxml import etree
 
-from .base import PullSession
+from .base import HttpSession
 
 
-def pull_plos(source: dict, sink: str) -> str:
+def pull_plos(source: dict, project: str, path: str) -> List[str]:
     """
     Pull a plos article.
 
@@ -34,22 +35,20 @@ def pull_plos(source: dict, sink: str) -> str:
 
     code = m.group(1)
     journal = journals[code]
-    article = m.group(2)
 
     url = "https://journals.plos.org/{}/article/file?id={}&type=manuscript".format(
         journal, doi
     )
 
-    folder, xml = os.path.split(sink)
+    folder, xml = os.path.split(path)
+    folder = os.path.join(project, folder)
 
-    assert "{}.{}.jats.xml".format(code, article) == xml
-
-    session = PullSession()
+    session = HttpSession()
     response = session.fetch_url(url)
     tree = etree.parse(io.BytesIO(response.content))
     root = tree.getroot()
     xlinkns = "http://www.w3.org/1999/xlink"
-    os.makedirs("{}.media".format(sink), exist_ok=True)
+    os.makedirs(os.path.join(project, "{}.media".format(path)), exist_ok=True)
 
     # Get the figures and rewrite hrefs
     for graphic in root.iterdescendants(tag="graphic"):
@@ -69,8 +68,8 @@ def pull_plos(source: dict, sink: str) -> str:
         new_href = "{}.media/{}".format(xml, filename)
         graphic.attrib["{%s}href" % xlinkns] = new_href
         graphic.attrib["mime-subtype"] = "png"
-        session.pull(url, "{}/{}".format(folder, new_href))
+        session.pull(url, os.path.join(folder, new_href))
 
-    tree.write(open(sink, "wb"))
+    tree.write(open(os.path.join(folder, xml), "wb"))
 
-    return sink
+    return [path]